+++ tests/testthat/test-hba--shiny.R
@@ -1,7 +1,8 @@
-# test_that("", {})
+test_that("UI succeeds", {
+  expect_no_error(hbaUI("ui"))
+})
 
 args <- list(simulation = shiny::reactive(test_sim_small()))
-
 shiny::testServer(hbaServer, args = args, {
   # Test that server returns an object of class erahumed_simulation()
   expect_no_error(session$returned())