+++ tests/testthat/test-shiny--main.R
@@ -1,3 +1,7 @@
+test_that("UI succeeds", {
+  expect_no_error(shiny_ui())
+})
+
 shiny::testServer(shiny_server, {
   expect_no_error(simulation())
 }) |> suppressMessages()