+++ NEWS.md
@@ -9,6 +9,7 @@
 
 * Added snapshot tests for plot methods of ERAHUMED layers (#149).
 * Added basic server tests for all shiny modules (#150).
+* Added basic UI no-error tests for all shiny modules.
 * Testing corner cases of assertions.
 
 # erahumed 0.11.2